+++ docs/whatsnew.rst
@@ -4,6 +4,8 @@ Whatsnew
 Version 0.3.4
 `````````````
 
+- API: Exposed the ``no-self-contained`` command-line option to the stitching
+  operation.
 - API: Added a ``warning`` option for controling whether stderr is included in the output.
 - API: Changed the ``on_error`` option to ``error`` for compatability with knitr and symmytry with the ``warning`` option.